refactor: first pass pickup-dispatch separation
