+++ src/components/kiosk/children/foreign/common/searchLoading.tsx
@@ -1,0 +1,7 @@
+export default function SearchLoading() {
+    return (
+        <div className="flex items-center justify-center w-full h-full">
+            <p className="text-gray-400 self-center">Loading...</p>
+        </div>
+    )
+}
